requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-property==0.2.2
attrs==25.3.0
Authlib==1.3.1
bcrypt==4.0.1
blinker==1.9.0
cached-property==1.5.2
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
Django==5.2
dotenv==0.9.9
ecdsa==0.19.1
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.1.24
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.70.0
grpcio-health-checking==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.26.0
huggingface-hub==0.30.2
idna==2.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jwcrypto==1.5.6
keras==3.8.0
keycloak==3.1.5
kiwisolver==1.4.8
langdetect==1.0.9
libclang==18.1.1
litellm==1.66.2
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.6.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.4.3
namex==0.0.8
narwhals==1.36.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.0.2
openai==1.74.0
openai-whisper==20240930
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pickle-mixin==1.0.2
pillow==11.1.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==5.29.3
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.4.8
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-jose==3.4.0
python-keycloak==5.5.0
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.24.0
rsa==4.9.1
rubicon-objc==0.5.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.61.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.1
validators==0.34.0
wcwidth==0.2.13
weaviate-client==4.11.0
websocket-client==1.8.0
websocket-server==0.6.4
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
